Some more refactoring on assets and renderers to better differentiate the purposes of each thing.
