+++ src/client/systems/views/starsystemview.h
@@ -27,7 +27,7 @@
 
 #include "client/systems/sysplanetterraingenerator.h"
 #include "common/universe.h"
-#include "engine/renderer/renderable.h"
+#include "engine/graphics/renderable.h"
 #include "engine/renderer/texturerenderer.h"
 #include "engine/renderer/renderer.h"
 #include "engine/application.h"
