+++ src/engine/asset/asset.h
