+++ src/engine/asset/assetmanager.cpp
@@ -14,7 +14,7 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "engine/assetmanager.h"
+#include "engine/asset/assetmanager.h"
 
 #include <spdlog/spdlog.h>
 #include <stb_image.h>
