+++ src/engine/graphics/primitives/polygon.cpp
