+++ src/engine/graphics/shader.cpp
@@ -14,7 +14,7 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "engine/renderer/shader.h"
+#include "engine/graphics/shader.h"
 
 #include <glad/glad.h>
