+++ src/engine/graphics/shader.h
@@ -22,7 +22,7 @@
 
 #include <glm/glm.hpp>
 
-#include "engine/asset.h"
+#include "engine/asset/asset.h"
 
 namespace cqsp {
 namespace asset {
