+++ src/engine/graphics/text.cpp
@@ -14,7 +14,7 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "engine/renderer/text.h"
+#include "engine/graphics/text.h"
 
 #include <spdlog/spdlog.h>
