+++ src/engine/graphics/texture.cpp
@@ -14,7 +14,7 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "engine/renderer/texture.h"
+#include "engine/graphics/texture.h"
 
 #define STB_IMAGE_IMPLEMENTATION
 #include <stb_image.h>
