+++ src/engine/graphics/texture.h
@@ -18,7 +18,7 @@
 
 #include <vector>
 
-#include "engine/asset.h"
+#include "engine/asset/asset.h"
 
 namespace cqsp {
 namespace asset {
